@@ -2,14 +2,14 @@ source 'https://rubygems.org'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '5.2.4'
+gem 'rails', '6.1.0'
 # Use mysql as the database for Active Record
 gem "mysql2"
 # Use postgresql as the database for Active Record
 # gem 'pg', '~> 0.15'
-gem 'puma', '~> 3.7'
+gem 'puma', '~> 4.1'
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
+gem 'sass-rails', '>= 6'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
@@ -20,7 +20,7 @@ gem 'coffee-rails', '~> 4.2'
 # Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
 gem 'turbolinks', '~> 5'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.5'
+gem 'jbuilder', '~> 2.7'
 # Use Redis adapter to run Action Cable in production
 # gem 'redis', '~> 3.0'
 
@@ -87,9 +87,6 @@ gem 'chartkick', '~> 3.0', '>= 3.0.1'
 # Session sotre
 gem 'activerecord-session_store'
 
-# 1.5.6: ruby >= 2.5.0
-gem 'dry-schema', '1.5.5'
-
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
@@ -97,7 +94,7 @@ group :development, :test do
   gem 'capybara', '~> 2.13'
   gem 'selenium-webdriver'
 
-  gem 'rspec-rails', '~> 3.8'
+  gem 'rspec-rails', '~> 4.0.1'
   gem 'spring-commands-rspec'
   gem 'factory_bot_rails'
   gem 'database_cleaner'
