@@ -331,7 +331,8 @@ class Teacher::Result::CompoundsController < ApplicationController
       if answer_score.nil?
         answer_score = AnswerScore.new(:page_id => @generic_page.id, :user_id => @user.id, answer_count: @count, :pass_cd => -1)
       elsif answer_score.answer_count != @count.to_i
-        answer_score = AnswerScoreHistory.where(:page_id => @generic_page.id, :user_id => @user.id, :answer_count => @count).first
+        history = AnswerScoreHistory.where(:page_id => @generic_page.id, :user_id => @user.id, :answer_count => @count).first
+        answer_score = history.answer_score
       end
       answer_score.save(validate: false) if answer_score.new_record?
 
@@ -340,11 +341,13 @@ class Teacher::Result::CompoundsController < ApplicationController
         if params[:score]
           params[:score].each do |question_id, score|
             ## 各解答に点数をセットしupdateする
-            answer = Answer.where(:question_id => question_id, :user_id => @user.id, :answer_count => @count).first
-            if answer.nil?
-              answer = Answer.new(:question_id => question_id, :user_id => @user.id, :answer_count => @count, :select_answer_id => 0, answer_score_id: answer_score.id)
-            elsif answer.answer_count != @count.to_i
+            if answer_score.answer_count != @count.to_i
               answer = AnswerHistory.where(:question_id => question_id, :user_id => @user.id, :answer_count => @count).first
+            else
+              answer = Answer.where(:question_id => question_id, :user_id => @user.id, :answer_count => @count).first
+              if answer.nil?
+                answer = Answer.new(:question_id => question_id, :user_id => @user.id, :answer_count => @count, :select_answer_id => 0, answer_score_id: answer_score.id)
+              end
             end
             answer.score = score
             answer.save
