@@ -29,7 +29,7 @@ module CustomValidationModule
   def validate_presence(target, message)
     value = self.send target.to_sym
     #errors.add(target.to_sym, message) if value.blank?
-    errors[:base] << message if value.blank?
+    errors.add(:base, message) if value.blank?
   end
 
   def validate_length(target, message, min, max = nil)
@@ -37,10 +37,10 @@ module CustomValidationModule
     unless value.blank?
       if max.nil?
         #errors.add(target.to_sym, message) if value.length < min
-        errors[:base] << message if value.length < min
+        errors.add(:base, message) if value.length < min
       else
         #errors.add(target.to_sym, message) if value.length < min || value.length > max
-        errors[:base] << message if value.length < min || value.length > max
+        errors.add(:base, message) if value.length < min || value.length > max
       end
     end
   end
@@ -49,7 +49,7 @@ module CustomValidationModule
     value = self.send target.to_sym
     unless value.blank?
       #errors.add(target.to_sym, message) if value.length > length
-      errors[:base] << message if value.length > length
+      errors.add(:base, message) if value.length > length
     end
   end
 
@@ -58,7 +58,7 @@ module CustomValidationModule
     value = self.send target.to_sym
     unless value.blank?
       #errors.add(target.to_sym, message) if value =~ valid_password
-      errors[:base] << message if value =~ valid_password
+      errors.add(:base, message) if value =~ valid_password
     end
   end
 
@@ -67,7 +67,7 @@ module CustomValidationModule
     value = self.send target.to_sym
     unless value.blank?
       #errors.add(target.to_sym, message) if value =~ valid_account
-      errors[:base] << message if value =~ valid_account
+      errors.add(:base, message) if value =~ valid_account
     end
   end
 
@@ -76,7 +76,7 @@ module CustomValidationModule
     value = self.send target.to_sym
     unless value.blank?
       #errors.add(target.to_sym, message) if value =~ valid_name
-      errors[:base] << message if value =~ valid_name
+      errors.add(:base, message) if value =~ valid_name
     end
   end
 
@@ -85,7 +85,7 @@ module CustomValidationModule
     value = self.send target.to_sym
     unless value.blank?
       #errors.add(target.to_sym, message) if value =~ valid_name
-      self.errors[:base] << message if value =~ valid_name
+      self.errors.add(:base, message) if value =~ valid_name
     end
   end
 
@@ -94,7 +94,7 @@ module CustomValidationModule
     value = self.send target.to_sym
     unless value.blank?
       #errors.add(target.to_sym, message) if value =~ valid_value
-      errors[:base] << message if value =~ valid_value
+      errors.add(:base, message) if value =~ valid_value
     end
   end
 
@@ -103,7 +103,7 @@ module CustomValidationModule
     value = self.send target.to_sym
     unless value.blank?
       #errors.add(target.to_sym, message) if value =~ valid_value
-      errors[:base] << message if value =~ valid_value
+      errors.add(:base, message) if value =~ valid_value
     end
   end
 
@@ -116,14 +116,14 @@ module CustomValidationModule
         count = self.class.where("#{target.to_s} = ? AND id != ?", value, self.id).count
       end
       #errors.add(target.to_sym, message) if count != 0
-      errors[:base] << message if count != 0
+      errors.add(:base, message) if count != 0
     end
   end
 
   def validate_mail_address(target, message)
     value = self.send target.to_sym
     #errors.add(target.to_sym, message) unless value =~ VALID_ADDRESS
-    errors[:base] << message unless value =~ VALID_ADDRESS
+    errors.add(:base, message) unless value =~ VALID_ADDRESS
   end
 
   def validate_date(target, message)
@@ -133,7 +133,7 @@ module CustomValidationModule
         date = value.to_date
       rescue ArgumentError
         #errors.add(target.to_sym, message)
-        errors[:base] << message
+        errors.add(:base, message)
       end
     end
   end
