@@ -1,6 +1,8 @@
 require 'rails_helper'
 
 RSpec.describe Compound, type: :model do
+  let(:test_file) { Rack::Test::UploadedFile.new(Rails.root.join('spec', 'fixtures', 'files', 'test.txt'), 'text/txt') }
+
   before do
     User.current_user = create(:teacher_user)
     @course = create(:course, :year_of_2019)
@@ -28,7 +30,7 @@ RSpec.describe Compound, type: :model do
       @compound.valid?
       expect(@compound.errors.count).to eq 0
 
-      @compound.file = fixture_file_upload('test.txt', 'text/txt')
+      @compound.file = test_file
       @compound.valid?
       expect(@compound.errors.count).to eq 0
     end
@@ -89,7 +91,7 @@ RSpec.describe Compound, type: :model do
   describe '登録' do
     it "ファイルアップロード成功" do
       file_name = 'test.txt'
-      compound = build(:compound, course_id: @course.id, file: fixture_file_upload('test.txt', 'text/txt'),
+      compound = build(:compound, course_id: @course.id, file: test_file,
         upload_flag: GenericPage::TYPE_FILEUPLOAD)
 
       expect{
@@ -103,10 +105,10 @@ RSpec.describe Compound, type: :model do
   describe '更新' do
     it "ファイルアップロード成功" do
       file_name = 'test_update.txt'
-      compound = create(:compound, course_id: @course.id, file: fixture_file_upload('test.txt', 'text/txt'),
+      compound = create(:compound, course_id: @course.id, file: test_file,
         upload_flag: GenericPage::TYPE_FILEUPLOAD)
 
-        compound.file = fixture_file_upload(file_name, 'text/txt')
+        compound.file = Rack::Test::UploadedFile.new(Rails.root.join('spec', 'fixtures', 'files', file_name), 'text/txt')
       expect{
         compound.save
       }.to change(GenericPage, :count).by(0)
@@ -117,7 +119,7 @@ RSpec.describe Compound, type: :model do
 
   describe '他コースへコピー' do
     before do
-      @compound = create(:compound, course_id: @course.id, file: fixture_file_upload('test.txt', 'text/txt'),
+      @compound = create(:compound, course_id: @course.id, file: test_file,
         upload_flag: GenericPage::TYPE_FILEUPLOAD,
         max_count: 2,
         pass_grade: 70,
@@ -155,7 +157,7 @@ RSpec.describe Compound, type: :model do
 
   describe '他テストからの設問コピー' do
     before do
-      @compound = create(:compound, course_id: @course.id, file: fixture_file_upload('test.txt', 'text/txt'),
+      @compound = create(:compound, course_id: @course.id, file: test_file,
         upload_flag: GenericPage::TYPE_FILEUPLOAD,
         max_count: 2,
         pass_grade: 70,
@@ -166,7 +168,7 @@ RSpec.describe Compound, type: :model do
         self_pass: "pass67890"
       )
 
-      @src_compound = create(:compound, course_id: @course.id, file: fixture_file_upload('test.txt', 'text/txt'),
+      @src_compound = create(:compound, course_id: @course.id, file: test_file,
         upload_flag: GenericPage::TYPE_FILEUPLOAD,
         max_count: 1,
         pass_grade: 65,
