@@ -2,5 +2,5 @@ require 'devise'
 
 RSpec.configure do |config|
   config.include Devise::Test::ControllerHelpers, type: :controller
-  config.include DeviceLoginHelper, :type => :controller
+  config.include DeviceLoginHelper
 end
